+++ AU/Private/AUPackage.ps1
@@ -36,4 +36,22 @@ class AUPackage {
         $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding($False)
         [System.IO.File]::WriteAllText($this.NuspecPath, $this.NuspecXml.InnerXml, $Utf8NoBomEncoding)
     }
+
+    Backup()  { 
+        $d = "$Env:TEMP\au\" + $this.Name
+
+        rm $d\* -Recurse -ea 0
+        cp . $d\_backup -Recurse 
+    }
+
+    [string] SaveAndRestore() { 
+        $d = "$Env:TEMP\au\" + $this.Name
+
+        cp . $d\_output -Recurse 
+        rm .\* -Recurse
+        cp $d\_backup\* . -Recurse 
+        
+        return "$d\_output"
+    }
+
 }